nang backend plumbing for onboarding (#100)
* fix aider in quick pick (#89)

* fix aider in quick pick

* remove console

* Added hot reload and cleaned state for aider (#90)

* Added hot reload and fixed state

* Added return type

---------


* Aider context providers fix in overlay (#92)

* added pearai_aider needlessly

* prrrroogress

* Added working

* Added working

* Back to pearai_model for aider

* Added cleanup

* Fixed bug

---------


* uiux improvements (#88)

* show loading on quickinput

* remove add more context providers from creator mode

* perplexity style streaming (working in sidebar)

* working perplexity like streaming in perplexity and aider

* rerenders and jump fixes after adding fade in perplexity style streaming

* fade in for headings, ul, ol, li (lists)

* remove add more context providers from perplexity

---------


* Added exclusions  (#93)

* Added exclusions

* Added exclusions

* Added exclusions

---------


* Added fix (#94)


* Aider fix (#96)

* Added progress

* Added restart button to aider

---------


* Aider fix 2 (#97)

* Added progress

* Added restart button to aider

* Added restart button to aider

---------


* Added provider name for metrics (#98)


* Added progress

* Added refactor

* Added progress

* Added progress

* Revert "Merge branch 'main' of https://github.com/trypear/pearai-submodule into nang/onboarding-hello"

This reverts commit 4dcd7c3051076cf47342a229148bfecada5b3469, reversing
changes made to 41319e51c2b8d5a47254e8062420267638ccc2c1.

* Actually working now

* Progress, sign in is messed up
'

* Added parity

* Added add to path

* Added path functionality

* Added better wording

* Added login cleanup

* Added login cleanup

* Fixed login trigger

* Cleaned up

* Cleaned up

---------
