+++ core/protocol/webview.ts
@@ -2,7 +2,6 @@ import type { ContextItemWithId, IndexingProgressUpdate } from "../index.js";
 import type { AiderStatusUpdate } from "../llm/llms/Aider";
 
 export type ToWebviewFromIdeOrCoreProtocol = {
-  addPearAIModel: [undefined, void];
   configUpdate: [undefined, void];
   getDefaultModelTitle: [undefined, string];
   loadMostRecentChat: [undefined, void];
@@ -10,6 +9,7 @@ export type ToWebviewFromIdeOrCoreProtocol = {
   aiderProcessStateUpdate: [AiderStatusUpdate, void];
   refreshSubmenuItems: [undefined, void];
   isContinueInputFocused: [undefined, boolean];
+  pearAISignedIn: [undefined, void];
   addContextItem: [
     {
       historyIndex: number;
