+++ extensions/intellij/src/main/kotlin/com/github/continuedev/continueintellijextension/toolWindow/ContinueBrowser.kt
@@ -60,7 +60,7 @@ class ContinueBrowser(val project: Project, url: String, useOsr: Boolean = false
         "llm/complete",
         "llm/streamComplete",
         "llm/streamChat",
-        "llm/resetPearAICredentials",
+        "llm/setPearAICredentials",
         "llm/listModels",
         "streamDiffLines",
         "stats/getTokensPerDay",
