+++ gui/package.json
@@ -45,6 +45,7 @@
     "posthog-js": "^1.130.1",
     "prismjs": "^1.29.0",
     "react": "^18.2.0",
+    "react-collapse": "^5.1.1",
     "react-dom": "^18.2.0",
     "react-error-boundary": "^4.0.11",
     "react-hook-form": "^7.47.0",
