+++ gui/src/App.tsx
@@ -22,7 +22,7 @@ import Stats from "./pages/stats";
 import Inventory from "./pages/inventory";
 import AiderGUI from "./integrations/aider/aidergui";
 import PerplexityGUI from "./integrations/perplexity/perplexitygui";
-import Welcome from "./pages/welcomeGui";
+import Welcome from "./pages/welcome/welcomeGui";
 
 
 declare global {
@@ -120,6 +120,8 @@ const router = createMemoryRouter(
   // TODO: Remove replace /welcome with /inventory when done testing
   {
     initialEntries: [window.isPearOverlay ? "/inventory" : window.initialRoute], 
+    // FOR DEV'ing welcome: 
+    // initialEntries: [window.isPearOverlay ? "/welcome" : window.initialRoute],
   },
 );
